requirements.txt
openai>=1.30.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
python-dotenv>=1.0.0
